requirements.txt
torch
torchvision
pillow
scikit-learn
streamlit
matplotlib